+++ web/__init__.py
@@ -5,16 +5,16 @@ from web.config import Config
 db = SQLAlchemy()
 
 
-def create_app(config_instance=None):
+def create_app(config_class=Config):
     app = Flask(__name__)
 
-    if not config_instance:
-        config_instance = Config()
+    config_instance = config_class()
     app.config.from_object(config_instance)
 
     db.init_app(app)
 
     from web.main import bp as main_bp
+
     app.register_blueprint(main_bp)
 
     return app